straming bot/requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
aiohttp==3.9.1
jinja2==3.1.2
aiofiles==23.2.1
python-dotenv==1.0.0
uvloop==0.19.0 ; sys_platform != 'win32'